refactor(suite): ConfirmOnDevice -> ConfirmOnDevicePill
